# docs/openapi.json
{
  "components": {
...
    "securitySchemes": {
      "BearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "EntityBearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "NoticeBearerAuth": {
        "scheme": "bearer",
        "type": "http"
      },
      "SelfBearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The subscription engine, extracted from Minty (Part 2 of the modernisation plan).",
    "title": "Minty Billing API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/onboarding/billing/accounts": {
      "get": {
        "description": "GET -> ``{accounts: [{id, billing_email, billing_company, default_id, cards[]}], …}``.\nPOST ``{payment_method, billing_email?, billing_company?}`` opens an account on a card the\npayer already holds (ownership proven first); it authorises nothing.",
        "operationId": "billing_api_onboarding_billing_accounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "The payer's billing accounts; open one",
        "tags": [
          "Onboarding"
        ]
      },
      "post": {
        "description": "GET -> ``{accounts: [{id, billing_email, billing_company, default_id, cards[]}], …}``.\nPOST ``{payment_method, billing_email?, billing_company?}`` opens an account on a card the\npayer already holds (ownership proven first); it authorises nothing.",
        "operationId": "billing_api_onboarding_billing_accounts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "The payer's billing accounts; open one",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/api/onboarding/billing/authorize": {
      "post": {
        "description": "Body ``{entity_id, payment_method?}`` -> ``{\"has_billing_consent\": true}``. Charges\nnothing: the trial runs its term, and this is the difference between \"converts\" and\n\"expires\" at the end of it. ``payment_method`` is nominated BEFORE consent is recorded\n(``establish_payer=True`` - the one caller that does, because during the wizard no\ncompany has a payer yet and this request is what establishes one). Idempotent.",
        "operationId": "billing_api_onboarding_billing_authorize",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Record consent to bill this company",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/api/onboarding/billing/payment-methods": {
      "get": {
        "operationId": "billing_api_onboarding_billing_payment_methods",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Every card on the payer's account",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/api/onboarding/billing/payment-methods/confirm": {
      "post": {
        "description": "Body ``{setup_intent, make_default?, billing_group_id?, billing_email?,\nbilling_company?}``. Saving a card AUTHORISES NOTHING - that is ``/billing/authorize``.\nThe three account fields are optional: a group id puts the card on an account the payer\nalready has (the service checks it is theirs), an email or company opens a new one.",
        "operationId": "billing_api_onboarding_billing_confirm",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Adopt the confirmed card",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/api/onboarding/billing/payment-methods/default": {
      "post": {
        "description": "Body ``{payment_method}``. Nominates nothing by itself; decides which card is offered\nfirst - including to ``/billing/authorize``.",
        "operationId": "billing_api_onboarding_billing_set_default",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Make one card the account's main one",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/api/onboarding/billing/payment-methods/setup-intent": {
      "post": {
        "description": "``{client_secret, publishable_key, setup_intent}``; needs no customer yet.",
        "operationId": "billing_api_onboarding_billing_setup_intent",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Open a SetupIntent for the card form",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/api/onboarding/payment-method": {
      "get": {
        "description": "``?entity_id=`` -> ``{\"has_payment_method\", \"has_billing_consent\", \"card\"}``.\n\nTwo separate questions, and the second is the one that matters: ``has_payment_method``\nis read live from Stripe and belongs to the PAYER, shared across every entity they pay\nfor; consent is per (entity, payer) and decides whether the 30-day trial converts or\nlapses. ``card`` IS THIS ENTITY'S NOMINATED CARD, not the payer's default - null when\nnothing is nominated, and null is the honest answer. Neither gates the wizard. Stripe\nunreachable degrades to \"no card\" rather than 500; the consent read is local and\ndeliberately outside that guard.",
        "operationId": "billing_api_onboarding_payment_method_status",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Whether this company can be billed",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/api/onboarding/payment-method/complete": {
      "post": {
        "description": "Body ``{entity_id, session_id}`` -> ``{\"has_payment_method\": true}``. Idempotent, so a\nrefresh of the return URL is harmless.",
        "operationId": "billing_api_onboarding_payment_method_complete",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Save the card a setup Checkout captured",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/api/onboarding/payment-method/setup": {
      "post": {
        "description": "Body ``{entity_id}`` -> ``{\"url\"}``. No charge, no subscription: onboarding starts the\ntrials at finalize. Stripe returns to the wizard with ``?pm_session_id=``, which it hands\nto ``/payment-method/complete``.",
        "operationId": "billing_api_onboarding_payment_method_setup",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Open a setup-mode Checkout to save a card",
        "tags": [
          "Onboarding"
        ]
      }
    },
    "/api/onboarding/trials/start": {
      "post": {
        "description": "Body ``{entity_id}`` -> ``{\"trial_end\": iso8601 | null}``.\n\nWhat Flask's ``onboarding_finalize`` did inline (``start_trials_for_enabled_modules``,\nbest-effort, then the earliest ``trial_end`` read back) - except that here a failure\nFAILS: onboarding-backend's finalize has already flipped the company live when it calls\nthis, and a trial that silently did not start is a company with modules on and nothing\nbehind them. A ``CheckoutError`` answers with its own status, anything else 502; the All\nSet screen offers Try again. Idempotent: modules that already hold a trial or a\nsubscription are skipped by the starter, and ``trial_end`` is READ BACK from the rows -\nthe answer on a revisit as much as on the first call. Null is a legitimate answer (an\nentity with no enabled module has no trial to state).",
        "operationId": "billing_api_onboarding_trials_start",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Start the card-free trials of a finished wizard",
        "tags": [
          "Onboarding"
        ]
      }
    }
  },
  "servers": []
}